+++ packages/kit/src/types/internal.d.ts
@@ -61,6 +61,7 @@ export interface AssetDependencies {
 	imports: string[];
 	stylesheets: string[];
 	fonts: string[];
+	stylesheet_map: Map<string, { css: Set<string>; assets: Set<string> }>;
 }
 
 export interface BuildData {